agendeja_rest/servicos/serializers.py
from rest_framework import serializers
from .models import Servico, Cliente

class ServicoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Servico
        fields = '__all__'

class ClienteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cliente
        fields = '__all__'
